Ai Generation coding , n8n,RabiitMQ

+++ src/main/java/com/aiatelye/leather/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package com.aiatelye.leather.repository;
 
-import com.aiatelye.leather.dao.ShippingLocation;
 import com.aiatelye.leather.dao.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
